KeiSuzuki/requirements.txt
ca-certificates=2019.1.23
certifi=2019.3.9
click=7.0
flask=1.0.2
itsdangerous=1.1.0
jinja2=2.10.1
libprotobuf=3.6.0
markupsafe=1.1.1
mysql-connector-c=6.1.11
mysql-connector-python=8.0.13
openssl=1.1.1b
pip=19.1.1
protobuf=3.6.0
python=3.7.3
setuptools=41.0.1
six=1.12.0
sqlite=3.28.0
vc=14.1
vs2015_runtime=14.15.26706
werkzeug=0.15.2
wheel=0.33.2
wincertstore=0.2
zlib=1.2.11
